+++ Asset Processors/Import Tasks/BaseImportTask.cs
@@ -12,6 +12,8 @@ namespace AssetTools
 	[System.Serializable]
 	public abstract class BaseImportTask : ScriptableObject, IImportTask
 	{
+		public enum ProcessingType { Pre, Post }
+		
 		protected List<string> m_AssetsToForceApply = new List<string>();
 		protected string m_SearchFilter = "";
 		
@@ -28,6 +30,8 @@ namespace AssetTools
 		}
 		
 		public abstract string AssetMenuFixString { get; }
+
+		public abstract ProcessingType TaskProcessType { get; }
 
 		public abstract bool CanProcess( AssetImporter item );
 
@@ -76,11 +80,11 @@ namespace AssetTools
 			return true;
 		}
 
-		public virtual bool Apply( AssetImporter importer, ImportDefinitionProfile fromProfile )
+		public virtual bool Apply( ImportContext context, ImportDefinitionProfile fromProfile )
 		{
-			if( CanProcess( importer ) == false )
+			if( CanProcess( context.Importer ) == false )
 				return false;
-			m_AssetsToForceApply.Remove( importer.assetPath );
+			m_AssetsToForceApply.Remove( context.Importer.assetPath );
 			return true;
 		}
 	}